WIP: monster check-in for new IR infrastructure and file I/O based test matrices

+++ include/universal/native/integers.hpp
@@ -173,7 +173,7 @@ inline std::string to_hex(const Integer& number, bool nibbleMarker = false, bool
 	uint64_t nibble{ 0 };
 	unsigned nibbleIndex = nbits / 4 - 1u;
 	unsigned rightShift = nibbleIndex * 4;
-	for (int i = nbits - 1; i >= 0; --i) {
+	for (int i = static_cast<int>(nbits - 1u); i >= 0; --i) {
 		nibble |= (mask & number);
 		mask >>= 1;
 		if ((i % 4) == 0) {
